4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -4,6 +4,19 @@ const Blog = require('../models/blog');
 const User = require('../models/user');
 
 
+const getUserFromToken = async (token, next) => {
+    const { id: uid = null } = jwt.verify(token, process.env.SECRET);
+
+    if (!uid)
+        return next({
+            name: 'AuthErr',
+            message: 'invalid token'
+        });
+
+    return await User.findById(uid);
+};
+
+
 router.get('/', async (_req, resp) => {
     const blogs = await Blog.find({}).populate('user', {
         username: 1, name: 1, _id: 1
@@ -14,15 +27,8 @@ router.get('/', async (_req, resp) => {
 
 
 router.post('/', async ({ token, body }, resp, next) => {
-    const { id: uid = null } = jwt.verify(token, process.env.SECRET);
 
-    if (!uid)
-        return next({
-            name: 'AuthErr',
-            message: 'invalid token'
-        });
-
-    const user = await User.findById(uid);
+    const user = getUserFromToken(token, next);
 
     const new_blog = new Blog({
         ...body,
@@ -40,8 +46,22 @@ router.post('/', async ({ token, body }, resp, next) => {
 });
 
 
-router.delete('/:id', async ({ params: { id } }, resp) => {
-    await Blog.deleteOne({ _id: id });
+router.delete('/:id', async ({ params: { id }, token }, resp, next) => {
+    const user = await getUserFromToken(token, next);
+
+    const blog = await Blog.findById(id);
+
+    if (blog.user._id.toString() !== user._id.toString())
+        return next({
+            name: 'AuthErr',
+            message: 'you SHALL NOT DELETE!'
+        });
+
+    user.blogs = user.blogs.filter(bid => bid !== blog._id);
+    await Promise.all([
+        user.save(),
+        Blog.deleteOne({ _id: id })
+    ]);
 
     resp.status(204).end();
 });
